cleaning up stores, refac ImageScaleUtil for usability

+++ src/main/java/org/morriswa/messageboard/util/CustomS3UtilImpl.java
@@ -13,15 +13,19 @@ import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.FileSystemException;
 import java.time.Instant;
 import java.util.Date;
+import java.util.UUID;
 
 @Component @Slf4j
 public class CustomS3UtilImpl implements CustomS3Util {
+    private final String INTERNAL_FILE_CACHE_PATH;
     private final Environment e;
     private final AmazonS3 s3;
     private final String ACTIVE_BUCKET;
@@ -31,15 +35,28 @@ public class CustomS3UtilImpl implements CustomS3Util {
         this.e = e;
         this.s3 = AmazonS3ClientBuilder.standard().withRegion(Regions.US_EAST_1).build();
         this.ACTIVE_BUCKET = e.getRequiredProperty("aws.s3.bucket");
+        this.INTERNAL_FILE_CACHE_PATH = e.getRequiredProperty("server.filecache");
     }
 
-    public void uploadObjectToS3(File toUpload, String destination) throws IOException {
+    public void uploadToS3(BufferedImage toUpload, String destination) throws IOException {
+
+        final UUID newImagePath = UUID.randomUUID();
+
+        File temp = new File(this.INTERNAL_FILE_CACHE_PATH + newImagePath);
+
+        ImageIO.write(toUpload, "png", temp);
+
+        if (!temp.exists()) {
+            throw new IOException("Bad");
+//                    String.format(e.getRequiredProperty("user-profile.service.errors.bad-image-format"),
+//                            imageRequest.getImageFormat()));
+        }
 
         s3.putObject(new PutObjectRequest(ACTIVE_BUCKET,
                 destination,
-                toUpload).withCannedAcl(CannedAccessControlList.Private));
+                temp).withCannedAcl(CannedAccessControlList.Private));
 
-        if (!toUpload.delete()) {
+        if (!temp.delete()) {
             throw new FileSystemException(
                     e.getRequiredProperty("user-profile.service.errors.unable-to-delete-cached-file"));
         }
